requirements.txt
fastapi
uvicorn
pymongo
requests
python-dotenv
tenacity
dnspython
dnspython>=3.0.0
